requirements.txt
selenium
playwright
openai
pyyaml
beautifulsoup4
requests
lxml
openai